Move ShortenedUrls to rover package
